+++ domains/client/relayer/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "domain-client-message-relayer"
+version = "0.1.0"
+authors = ["Vedhavyas Singareddi <ved@subspace.network>"]
+edition = "2021"
+homepage = "https://subspace.network"
+repository = "https://github.com/subspace/subspace"
+description = "Subspace message relayer."
+include = [
+    "/src",
+    "/Cargo.toml",
+]
+
+[dependencies]
+# Substrate dependencies
+sc-network = { git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
+sc-network-gossip = { git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
+sc-utils = { git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
+sp-core = { git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
+sp-runtime = { git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
+
+futures = "0.3.21"
+parity-scale-codec = { version = "3.1.5", features = ["derive"] }
+parking_lot = "0.12.1"
+tracing = "0.1.36"
+
+sp-messenger = { path = "../../primitives/messenger" }
